+++ govtool/frontend/src/utils/bech32Validation.ts
@@ -7,7 +7,7 @@ import I18n from "@/i18n";
  * @param value - The value to be validated.
  * @returns A boolean indicating whether the value is valid or an error message if it is not valid.
  */
-export default async (value: string) => {
+export const bech32Validation = async (value: string) => {
   try {
     const decoded = await bech32.decode(value);
     if (decoded.words.length) {